[tutorials]Cleanup warnings. Now all tutorials are totally warnings free again (at least on win8 + VS2012).

+++ IndieLib/tutorials/advanced/01_IND_Surface_Grids/Tutorial01.cpp
@@ -147,7 +147,7 @@ Indielib_Main
 		mBack->setAngleXYZ(0, 0, mAngle);
 
 		// Update grid vertices for making a "wave" effect
-		mT = mTimer->getTicks() / 1000.0f;
+		mT = (float) mTimer->getTicks() / 1000.0f;
 		
 		for (int i = 1; i < mNumBlocksX; i++)
 			for (int j = 1; j < mNumBlocksY; j++)
